Reorganize ircnet module hierarchy
