+++ crates/ircnet/src/client/autoresponders/ctcp.rs
