+++ crates/ircnet/src/client/autoresponders/responderset.rs
