+++ crates/ircnet/src/connect/codecs/raw_message.rs
