+++ crates/ircnet/src/lib.rs
@@ -1,9 +1,2 @@
-pub mod autoresponders;
-mod client;
-pub mod codecs;
-pub mod commands;
-mod connect;
-mod consts;
-pub use crate::client::*;
-pub use crate::connect::*;
-pub use crate::consts::*;
+pub mod client;
+pub mod connect;
